requirements.txt
pandas == 1.4.2
streamlit == 1.40.1
streamlit-aggrid == 0.3.3
numpy == 1.26.4
psycopg2-binary
python-docx
openpyxl
fpdf
python-dotenv
llama-cpp-python == 0.3.9